requirements/local.txt
-r base.txt

nose==1.3.4
pep8==1.6.2
django-debug-toolbar==1.4
django-extensions==1.5.9
django-nose==1.3
